requirements.txt
# Raphael beta bundle — install: pip install -r requirements.txt
telethon>=1.34.0
pandas>=2.0.0
oandapyV20>=0.7.2
requests>=2.28.0
Pillow>=10.0.0
python-dotenv>=1.0.0
